chore: clean up duplicate llama code
